requirements.txt
 Django>=2.0
 psycopg2==2.7.3.2 
 djangorestframework
 django-rest-swagger
 arrow
